fetched data from rest api can be dispatched successfully now :)

+++ services/rest/masterdata/src/main/java/ruixue/rest/repository/ShopsRepository.java
@@ -1,10 +1,9 @@
 package ruixue.rest.repository;
 
-import java.math.BigInteger;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import ruixue.rest.entity.Shop;
 
-public interface ShopsRepository extends MongoRepository<Shop, BigInteger> {
+public interface ShopsRepository extends MongoRepository<Shop, Integer> {
 }
